+++ src/components/Page/index.tsx
@@ -1,2 +1,6 @@
 export { Page } from './Page';
-export type { FlowbitePageRootTheme, FlowbitePageTheme, PageColors, PageProps } from './Page';
+export type { 
+  FlowbitePageTheme, 
+  PageColors, 
+  PageProps,
+ } from './Page';