+++ src/components/Page/theme.ts
@@ -1,16 +1,16 @@
 import type { FlowbitePageTheme } from './Page';
 
 export const PageTheme: FlowbitePageTheme = {
-  root: {
-    base: 'text-sm font-medium',
-    disabled: 'opacity-50',
+    base: 'p-4',
     colors: {
       default: 'text-gray-900 dark:text-white',
-      info: 'text-cyan-500 dark:text-cyan-600',
+      primary:'',
+      secondary:'bg-secondary-100',
+      info: 'bg-cyan-500 dark:bg-cyan-600',
       failure: 'text-red-700 dark:text-red-500',
       warning: 'text-yellow-500 dark:text-yellow-600',
       success: 'text-green-700 dark:text-green-500',
     },
-    align: ''
-  },
+    align: '',
+    radius:''
 };