+++ src/components/Section/index.tsx
@@ -1,3 +1,10 @@
 export { Section } from './Section';
-export type { FlowbiteSectionRootTheme, FlowbiteSectionTheme, SectionColors, SectionProps,
-  FlowbiteSectionOutlineTheme,SectionBorderRadius } from './Section';
+export type { 
+  FlowbiteSectionRootTheme, 
+  FlowbiteSectionTheme, 
+  SectionColors, 
+  SectionProps,
+  FlowbiteSectionOutlineTheme,
+  SectionBorderRadius,
+  SectionOutlineColors,
+} from './Section';